+++ src/codex_agent.rs
@@ -1,13 +1,11 @@
 use agent_client_protocol::{
-    Agent, AgentCapabilities, AgentSideConnection, Annotations, AudioContent, AuthenticateRequest,
-    AuthenticateResponse, BlobResourceContents, CancelNotification, Client, ContentBlock, Diff,
-    EmbeddedResource, EmbeddedResourceResource, Error, ImageContent, InitializeRequest,
-    InitializeResponse, LoadSessionRequest, LoadSessionResponse, McpCapabilities, McpServer,
-    NewSessionRequest, NewSessionResponse, PermissionOption, PermissionOptionId,
-    PermissionOptionKind, PromptCapabilities, PromptRequest, PromptResponse,
-    RequestPermissionOutcome, RequestPermissionRequest, ResourceLink, SessionId,
-    SessionNotification, SessionUpdate, SetSessionModeRequest, SetSessionModeResponse,
-    SetSessionModelRequest, SetSessionModelResponse, TextContent, TextResourceContents, ToolCall,
+    Agent, AgentCapabilities, AgentSideConnection, AuthenticateRequest, AuthenticateResponse,
+    CancelNotification, Client, Diff, Error, InitializeRequest, InitializeResponse,
+    LoadSessionRequest, LoadSessionResponse, McpCapabilities, McpServer, NewSessionRequest,
+    NewSessionResponse, PermissionOption, PermissionOptionId, PermissionOptionKind,
+    PromptCapabilities, PromptRequest, PromptResponse, RequestPermissionOutcome,
+    RequestPermissionRequest, SessionId, SessionNotification, SessionUpdate, SetSessionModeRequest,
+    SetSessionModeResponse, SetSessionModelRequest, SetSessionModelResponse, ToolCall,
     ToolCallContent, ToolCallId, ToolCallLocation, ToolCallStatus, ToolCallUpdate,
     ToolCallUpdateFields, ToolKind, V1,
 };
@@ -18,13 +16,12 @@ use codex_core::{
     config::Config,
     config_types::{McpServerConfig, McpServerTransportConfig},
     protocol::{
-        ApplyPatchApprovalRequestEvent, FileChange, McpInvocation, PatchApplyBeginEvent,
-        PatchApplyEndEvent, ReviewDecision,
+        ApplyPatchApprovalRequestEvent, FileChange, PatchApplyBeginEvent, PatchApplyEndEvent,
+        ReviewDecision,
     },
 };
 use codex_protocol::ConversationId;
 use itertools::Itertools;
-use mcp_types::CallToolResult;
 use std::{
     cell::RefCell,
     collections::HashMap,
@@ -121,145 +118,6 @@ impl CodexAgent {
 
         if let Err(e) = self.client().session_notification(notification).await {
             error!("Failed to send session notification: {:?}", e);
-        }
-    }
-
-    async fn start_mcp_tool_call(
-        &self,
-        session_id: SessionId,
-        call_id: String,
-        invocation: McpInvocation,
-    ) {
-        // Create a ToolCall so subsequent ToolCallUpdate (e.g. terminal embedding) can attach by id.
-        let tool_call_id = ToolCallId(call_id.clone().into());
-        let title = format!("Tool: {}/{}", invocation.server, invocation.tool);
-        self.send_notification(
-            session_id.clone(),
-            SessionUpdate::ToolCall(ToolCall {
-                id: tool_call_id,
-                title,
-                kind: ToolKind::Other,
-                status: ToolCallStatus::InProgress,
-                content: vec![],
-                locations: vec![],
-                raw_input: Some(serde_json::json!(&invocation)),
-                raw_output: None,
-                meta: None,
-            }),
-        )
-        .await;
-    }
-
-    async fn end_mcp_tool_call(
-        &self,
-        session_id: SessionId,
-        call_id: String,
-        result: Result<CallToolResult, String>,
-    ) {
-        let is_error = match result.as_ref() {
-            Ok(result) => result.is_error.unwrap_or_default(),
-            Err(_) => true,
-        };
-        let raw_output = match result.as_ref() {
-            Ok(result) => serde_json::json!(result),
-            Err(err) => serde_json::json!(err),
-        };
-        self.send_notification(
-            session_id,
-            SessionUpdate::ToolCallUpdate(ToolCallUpdate {
-                id: ToolCallId(call_id.into()),
-                fields: ToolCallUpdateFields {
-                    status: Some(if is_error {
-                        ToolCallStatus::Failed
-                    } else {
-                        ToolCallStatus::Completed
-                    }),
-                    content: result.ok().filter(|result| !result.content.is_empty()).map(
-                        |result| {
-                            result
-                                .content
-                                .into_iter()
-                                .map(Self::codex_content_to_acp_content)
-                                .collect()
-                        },
-                    ),
-                    raw_output: Some(raw_output),
-                    ..Default::default()
-                },
-                meta: None,
-            }),
-        )
-        .await;
-    }
-
-    fn codex_content_to_acp_content(content: mcp_types::ContentBlock) -> ToolCallContent {
-        ToolCallContent::Content {
-            content: match content {
-                mcp_types::ContentBlock::TextContent(text_content) => {
-                    ContentBlock::Text(TextContent {
-                        annotations: text_content.annotations.map(convert_annotations),
-                        text: text_content.text,
-                        meta: None,
-                    })
-                }
-                mcp_types::ContentBlock::ImageContent(image_content) => {
-                    ContentBlock::Image(ImageContent {
-                        annotations: image_content.annotations.map(convert_annotations),
-                        data: image_content.data,
-                        mime_type: image_content.mime_type,
-                        uri: None,
-                        meta: None,
-                    })
-                }
-                mcp_types::ContentBlock::AudioContent(audio_content) => {
-                    ContentBlock::Audio(AudioContent {
-                        annotations: audio_content.annotations.map(convert_annotations),
-                        data: audio_content.data,
-                        mime_type: audio_content.mime_type,
-                        meta: None,
-                    })
-                }
-                mcp_types::ContentBlock::ResourceLink(resource_link) => {
-                    ContentBlock::ResourceLink(ResourceLink {
-                        annotations: resource_link.annotations.map(convert_annotations),
-                        description: resource_link.description,
-                        mime_type: resource_link.mime_type,
-                        name: resource_link.name,
-                        size: resource_link.size,
-                        title: resource_link.title,
-                        uri: resource_link.uri,
-                        meta: None,
-                    })
-                }
-                mcp_types::ContentBlock::EmbeddedResource(embedded_resource) => {
-                    ContentBlock::Resource(EmbeddedResource {
-                        annotations: embedded_resource.annotations.map(convert_annotations),
-                        resource: match embedded_resource.resource {
-                            mcp_types::EmbeddedResourceResource::TextResourceContents(
-                                text_resource_contents,
-                            ) => EmbeddedResourceResource::TextResourceContents(
-                                TextResourceContents {
-                                    mime_type: text_resource_contents.mime_type,
-                                    text: text_resource_contents.text,
-                                    uri: text_resource_contents.uri,
-                                    meta: None,
-                                },
-                            ),
-                            mcp_types::EmbeddedResourceResource::BlobResourceContents(
-                                blob_resource_contents,
-                            ) => EmbeddedResourceResource::BlobResourceContents(
-                                BlobResourceContents {
-                                    blob: blob_resource_contents.blob,
-                                    mime_type: blob_resource_contents.mime_type,
-                                    uri: blob_resource_contents.uri,
-                                    meta: None,
-                                },
-                            ),
-                        },
-                        meta: None,
-                    })
-                }
-            },
         }
     }
 
@@ -693,22 +551,5 @@ impl Agent for CodexAgent {
             .await?;
 
         Ok(SetSessionModelResponse::default())
-    }
-}
-
-fn convert_annotations(annotations: mcp_types::Annotations) -> Annotations {
-    Annotations {
-        audience: annotations.audience.map(|audience| {
-            audience
-                .into_iter()
-                .map(|audience| match audience {
-                    mcp_types::Role::Assistant => agent_client_protocol::Role::Assistant,
-                    mcp_types::Role::User => agent_client_protocol::Role::User,
-                })
-                .collect()
-        }),
-        last_modified: annotations.last_modified,
-        priority: annotations.priority,
-        meta: None,
     }
 }
