+++ src/conversation.rs
@@ -4,8 +4,9 @@ use std::{
 };
 
 use agent_client_protocol::{
-    AgentSideConnection, Client as _, ContentBlock, EmbeddedResource, EmbeddedResourceResource,
-    Error, LoadSessionResponse, ModelId, ModelInfo, PermissionOption, PermissionOptionId,
+    AgentSideConnection, Annotations, AudioContent, BlobResourceContents, Client as _,
+    ContentBlock, EmbeddedResource, EmbeddedResourceResource, Error, ImageContent,
+    LoadSessionResponse, ModelId, ModelInfo, PermissionOption, PermissionOptionId,
     PermissionOptionKind, Plan, PlanEntry, PlanEntryPriority, PlanEntryStatus, PromptRequest,
     RequestPermissionOutcome, RequestPermissionRequest, RequestPermissionResponse, ResourceLink,
     SessionId, SessionMode, SessionModeId, SessionModeState, SessionModelState,
@@ -24,11 +25,13 @@ use codex_core::{
     protocol::{
         AgentMessageDeltaEvent, AgentReasoningDeltaEvent, AgentReasoningRawContentDeltaEvent,
         AgentReasoningSectionBreakEvent, Event, EventMsg, ExecApprovalRequestEvent,
-        ExecCommandBeginEvent, ExecCommandEndEvent, ExecCommandOutputDeltaEvent, InputItem, Op,
-        ReviewDecision, TaskStartedEvent, UserMessageEvent, WebSearchBeginEvent, WebSearchEndEvent,
+        ExecCommandBeginEvent, ExecCommandEndEvent, ExecCommandOutputDeltaEvent, InputItem,
+        McpInvocation, McpToolCallBeginEvent, McpToolCallEndEvent, Op, ReviewDecision,
+        TaskStartedEvent, UserMessageEvent, WebSearchBeginEvent, WebSearchEndEvent,
     },
 };
 use codex_protocol::config_types::ReasoningEffort;
+use mcp_types::CallToolResult;
 use tokio::sync::{mpsc, oneshot};
 use tracing::{error, info};
 
@@ -322,14 +325,14 @@ impl PromptState {
                 );
                 self.exec_command_end(client, end_event).await;
             }
-            // EventMsg::McpToolCallBegin(McpToolCallBeginEvent { call_id, invocation }) => {
-            //     info!("MCP tool call begin: call_id={call_id}, invocation={} {}", invocation.server, invocation.tool);
-            //     self.start_mcp_tool_call(session_id.clone(), call_id, invocation).await;
-            // }
-            // EventMsg::McpToolCallEnd(McpToolCallEndEvent { call_id, invocation, duration, result }) => {
-            //     info!("MCP tool call ended: call_id={call_id}, invocation={} {}, duration={duration:?}", invocation.server, invocation.tool);
-            //     self.end_mcp_tool_call(session_id.clone(), call_id, result).await;
-            // }
+            EventMsg::McpToolCallBegin(McpToolCallBeginEvent { call_id, invocation }) => {
+                info!("MCP tool call begin: call_id={call_id}, invocation={} {}", invocation.server, invocation.tool);
+                self.start_mcp_tool_call(client, call_id, invocation).await;
+            }
+            EventMsg::McpToolCallEnd(McpToolCallEndEvent { call_id, invocation, duration, result }) => {
+                info!("MCP tool call ended: call_id={call_id}, invocation={} {}, duration={duration:?}", invocation.server, invocation.tool);
+                self.end_mcp_tool_call(client, call_id, result).await;
+            }
             // // File based events
             // EventMsg::ApplyPatchApprovalRequest(event) => {
             //     info!("Apply patch approval request: call_id={}, reason={:?}", event.call_id, event.reason);
@@ -393,6 +396,69 @@ impl PromptState {
                 todo!()
             }
         }
+    }
+
+    async fn start_mcp_tool_call(
+        &self,
+        client: &Client,
+        call_id: String,
+        invocation: McpInvocation,
+    ) {
+        let tool_call_id = ToolCallId(call_id.clone().into());
+        let title = format!("Tool: {}/{}", invocation.server, invocation.tool);
+        client
+            .send_notification(SessionUpdate::ToolCall(ToolCall {
+                id: tool_call_id,
+                title,
+                kind: ToolKind::Other,
+                status: ToolCallStatus::InProgress,
+                content: vec![],
+                locations: vec![],
+                raw_input: Some(serde_json::json!(&invocation)),
+                raw_output: None,
+                meta: None,
+            }))
+            .await;
+    }
+
+    async fn end_mcp_tool_call(
+        &self,
+        client: &Client,
+        call_id: String,
+        result: Result<CallToolResult, String>,
+    ) {
+        let is_error = match result.as_ref() {
+            Ok(result) => result.is_error.unwrap_or_default(),
+            Err(_) => true,
+        };
+        let raw_output = match result.as_ref() {
+            Ok(result) => serde_json::json!(result),
+            Err(err) => serde_json::json!(err),
+        };
+        client
+            .send_notification(SessionUpdate::ToolCallUpdate(ToolCallUpdate {
+                id: ToolCallId(call_id.into()),
+                fields: ToolCallUpdateFields {
+                    status: Some(if is_error {
+                        ToolCallStatus::Failed
+                    } else {
+                        ToolCallStatus::Completed
+                    }),
+                    content: result.ok().filter(|result| !result.content.is_empty()).map(
+                        |result| {
+                            result
+                                .content
+                                .into_iter()
+                                .map(codex_content_to_acp_content)
+                                .collect()
+                        },
+                    ),
+                    raw_output: Some(raw_output),
+                    ..Default::default()
+                },
+                meta: None,
+            }))
+            .await;
     }
 
     async fn exec_approval(
@@ -1082,5 +1148,87 @@ fn format_uri_as_link(name: Option<String>, uri: String) -> String {
         format!("[@{name}]({uri})")
     } else {
         uri
+    }
+}
+
+fn codex_content_to_acp_content(content: mcp_types::ContentBlock) -> ToolCallContent {
+    ToolCallContent::Content {
+        content: match content {
+            mcp_types::ContentBlock::TextContent(text_content) => ContentBlock::Text(TextContent {
+                annotations: text_content.annotations.map(convert_annotations),
+                text: text_content.text,
+                meta: None,
+            }),
+            mcp_types::ContentBlock::ImageContent(image_content) => {
+                ContentBlock::Image(ImageContent {
+                    annotations: image_content.annotations.map(convert_annotations),
+                    data: image_content.data,
+                    mime_type: image_content.mime_type,
+                    uri: None,
+                    meta: None,
+                })
+            }
+            mcp_types::ContentBlock::AudioContent(audio_content) => {
+                ContentBlock::Audio(AudioContent {
+                    annotations: audio_content.annotations.map(convert_annotations),
+                    data: audio_content.data,
+                    mime_type: audio_content.mime_type,
+                    meta: None,
+                })
+            }
+            mcp_types::ContentBlock::ResourceLink(resource_link) => {
+                ContentBlock::ResourceLink(ResourceLink {
+                    annotations: resource_link.annotations.map(convert_annotations),
+                    description: resource_link.description,
+                    mime_type: resource_link.mime_type,
+                    name: resource_link.name,
+                    size: resource_link.size,
+                    title: resource_link.title,
+                    uri: resource_link.uri,
+                    meta: None,
+                })
+            }
+            mcp_types::ContentBlock::EmbeddedResource(embedded_resource) => {
+                ContentBlock::Resource(EmbeddedResource {
+                    annotations: embedded_resource.annotations.map(convert_annotations),
+                    resource: match embedded_resource.resource {
+                        mcp_types::EmbeddedResourceResource::TextResourceContents(
+                            text_resource_contents,
+                        ) => EmbeddedResourceResource::TextResourceContents(TextResourceContents {
+                            mime_type: text_resource_contents.mime_type,
+                            text: text_resource_contents.text,
+                            uri: text_resource_contents.uri,
+                            meta: None,
+                        }),
+                        mcp_types::EmbeddedResourceResource::BlobResourceContents(
+                            blob_resource_contents,
+                        ) => EmbeddedResourceResource::BlobResourceContents(BlobResourceContents {
+                            blob: blob_resource_contents.blob,
+                            mime_type: blob_resource_contents.mime_type,
+                            uri: blob_resource_contents.uri,
+                            meta: None,
+                        }),
+                    },
+                    meta: None,
+                })
+            }
+        },
+    }
+}
+
+fn convert_annotations(annotations: mcp_types::Annotations) -> Annotations {
+    Annotations {
+        audience: annotations.audience.map(|audience| {
+            audience
+                .into_iter()
+                .map(|audience| match audience {
+                    mcp_types::Role::Assistant => agent_client_protocol::Role::Assistant,
+                    mcp_types::Role::User => agent_client_protocol::Role::User,
+                })
+                .collect()
+        }),
+        last_modified: annotations.last_modified,
+        priority: annotations.priority,
+        meta: None,
     }
 }
